Update delete and view

+++ src/main/java/seedu/address/logic/commands/DeleteCommand.java
@@ -12,7 +12,7 @@ import seedu.address.model.Model;
 import seedu.address.model.person.Person;
 
 /**
- * Deletes a person identified using it's displayed index from the address book.
+ * Deletes a person identified using its displayed index from the address book.
  */
 public class DeleteCommand extends Command {
 
@@ -41,6 +41,21 @@ public class DeleteCommand extends Command {
         }
 
         Person personToDelete = lastShownList.get(targetIndex.getZeroBased());
+
+        /*
+        this part of code should only exist in javafx file so remove it so far
+        Show confirmation dialog
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+        alert.setTitle("Delete Confirmation");
+        alert.setHeaderText("Are you sure you want to delete this person?");
+        alert.setContentText(String.format("Name: %s", personToDelete.getName()));
+
+        ButtonType result = alert.showAndWait().orElse(ButtonType.CANCEL);
+        if (result != ButtonType.OK) {
+            return new CommandResult("Deletion cancelled.");
+        }
+         */
+
         model.deletePerson(personToDelete);
         return new CommandResult(String.format(MESSAGE_DELETE_PERSON_SUCCESS, Messages.format(personToDelete)));
     }
